% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{align_fasta}
\alias{align_fasta}
\title{Read a fasta file and align the sequences}
\usage{
align_fasta(file)
}
\arguments{
\item{file}{a fasta file located in the data folder}
}
\value{
a sequence alignment (msa)
}
\description{
Read a fasta file and align the sequences
}
