% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fasta.R
\name{build_fasta}
\alias{build_fasta}
\title{build a fasta file}
\usage{
build_fasta(term, retmax = 500, rettype = "fasta")
}
\arguments{
\item{term}{output of cb_taxa_gene function, gives a list of taxon and the name of the gene}

\item{retmax}{max number of sequences to retain; default is 500}

\item{rettype}{format of the output; default is Fasta}
}
\value{
return a list of fasta sequences
}
\description{
Download sequences from Genbank and build a fasta file from an object output by cb_taxa_gene function
}
\examples{
pom <- cb_taxa_gene("Pomatoschistus adriaticus","COI")
seq_fasta <- build_fasta(pom)
}
