% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_MLtree.R
\name{build_MLtree}
\alias{build_MLtree}
\title{Build a ML tree}
\usage{
build_MLtree(fas_align, output_name)
}
\arguments{
\item{fas_align}{a fasta alignement output by align_fasta function (msa object)}

\item{output_name}{a string indicating the name of the .tre file saved in the output folder}
}
\value{
a tree object
}
\description{
Build a ML tree
}
