% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_taxa_gene.R
\name{cb_taxa_gene}
\alias{cb_taxa_gene}
\title{Concatenate taxa names and gene name to build a search term for entrez_search function}
\usage{
cb_taxa_gene(tax, gene)
}
\arguments{
\item{tax}{a char vector, use c() for multiple names}

\item{gene}{a char vector, currently accepts a single gene}
}
\value{
the search term that feeds the entrez_search function
}
\description{
takes one or multiple taxa namzs, concatenate them and with the name of the gene build a search term for entrez_search function
}
\examples{
cb_taxa_gene(c("Pomatoschistus","Gobius"),"COI")


}
