% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{disp_access}
\alias{disp_access}
\title{Display accession number and species name of a fasta file}
\usage{
disp_access(file)
}
\arguments{
\item{file}{an object output by function read_fasta}
}
\value{
display a list of accessions with the index
}
\description{
Display accession number and species name of a fasta file
}
