% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{fas_name}
\alias{fas_name}
\title{rename fasta object}
\usage{
fas_name(fasta_ls)
}
\arguments{
\item{fasta_ls}{fasta object}
}
\value{
the same fasta object but with new names
}
\description{
rename fasta object
}
