% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishphylo-package.R
\docType{package}
\name{fishphylo-package}
\alias{fishphylo}
\alias{fishphylo-package}
\title{fishphylo: Analyses of DNA Sequences}
\description{
Import sequences from genebank, read new new sequences from fasta files, align and calculate phylogenetic trees.
}
\author{
\strong{Maintainer}: Julien Renoult \email{julien.renoult@cefe.cnrs.fr}

}
\keyword{internal}
