% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{plot a phylogenetic tree}
\usage{
plot_tree(tree, root)
}
\arguments{
\item{tree}{object output by build_NJtree function}
}
\value{
plot a tree
}
\description{
plot a phylogenetic tree
}
