% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{read_fasta}
\alias{read_fasta}
\title{read fasta file from output directory}
\usage{
read_fasta(name)
}
\arguments{
\item{name}{name of output file}

\item{file}{.fas file located in the "R" directory}
}
\value{
an object reading fasta sequences
}
\description{
read fasta file from output directory
}
\examples{
fas_file <- read_fasta("bueniaETpomadri_COI_seqs.fasta")
}
