% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{write_fasta}
\alias{write_fasta}
\title{write fasta file in output directory}
\usage{
write_fasta(fasta_ls, name = "fasta_seq")
}
\arguments{
\item{fasta_ls}{list of fasta sequence produced by build_fasta function}

\item{name}{name of output file}
}
\value{
write a file with .fasta extension
}
\description{
write fasta file in output directory
}
\examples{
ls_tax_gen <- cb_taxa_gene("Pomatoschistus adriaticus","COI")
ls_fasta <-build_fasta(ls_tax_gen)
write_fasta(ls_fasta, "pomadri_COI_seqs.fasta")

}
